Moves tests for check_missing_pdfs into separate file
